Cosmetical change: renamed 'isoline' folder to 'contour'.
